backend/ingestion-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
aiohttp